Improve  req&reply handler

+++ EspNowAESBroadcast.h
@@ -17,13 +17,13 @@
     void espNowAESBroadcast_setToMasterRole(bool master=true, unsigned char ttl=0 /*ttl for sync messages*/);
     void espNowAESBroadcast_setToBatteryNode(bool isBatteryNode=true);
 
-    void espNowAESBroadcast_RecvCB(void (*callback)(const uint8_t *, int, uint16_t));
+    void espNowAESBroadcast_RecvCB(void (*callback)(const uint8_t *, int, uint32_t));
     bool espNowAESBroadcast_send(uint8_t* msg, int size, int ttl=0); //Max message length is 236byte
     void espNowAESBroadcast_secredkey(const unsigned char key[16]);
     void espNowAESBroadcast_setAesInitializationVector(const unsigned char iv[16]);
 
     bool espNowAESBroadcast_sendAndHandleReply(uint8_t* msg, int size, int ttl, void (*f)(const uint8_t *, int)); //Max message length is 236byte
-    bool espNowAESBroadcast_sendReply(uint8_t* msg, int size, int ttl, uint16_t replyIdentifier);
+    bool espNowAESBroadcast_sendReply(uint8_t* msg, int size, int ttl, uint32_t replyIdentifier);
 
 
     void espNowAESBroadcast_requestInstantTimeSyncFromMaster(); //Only battery devices should use this!!!
